backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
openai
python-multipart
httpx
aiohttp
